created a solution, h and main.cpp files, added the source file to the project, added the solution to task 1
