+++ Extra-task-1/Extra-task-1.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.9.34728.123
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Extra-task-1", "Extra-task-1\Extra-task-1.vcxproj", "{4741EEB5-6AA8-42D3-B29B-A6736E811AA4}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{4741EEB5-6AA8-42D3-B29B-A6736E811AA4}.Debug|x64.ActiveCfg = Debug|x64
+		{4741EEB5-6AA8-42D3-B29B-A6736E811AA4}.Debug|x64.Build.0 = Debug|x64
+		{4741EEB5-6AA8-42D3-B29B-A6736E811AA4}.Debug|x86.ActiveCfg = Debug|Win32
+		{4741EEB5-6AA8-42D3-B29B-A6736E811AA4}.Debug|x86.Build.0 = Debug|Win32
+		{4741EEB5-6AA8-42D3-B29B-A6736E811AA4}.Release|x64.ActiveCfg = Release|x64
+		{4741EEB5-6AA8-42D3-B29B-A6736E811AA4}.Release|x64.Build.0 = Release|x64
+		{4741EEB5-6AA8-42D3-B29B-A6736E811AA4}.Release|x86.ActiveCfg = Release|Win32
+		{4741EEB5-6AA8-42D3-B29B-A6736E811AA4}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {E6CE83C2-1751-4474-ADC7-5CC53FDA1E37}
+	EndGlobalSection
+EndGlobal
