+++ Extra-task-1/Extra-task-1/extra-task-1.h
@@ -1,0 +1,6 @@
+#ifndef EXTRA_TASK_1_H
+#define EXTRA_TASK_1_H
+
+double seconds_difference(double time_1, double time_2);
+#endif // !EXTRA_TASK_1_H
+
