+++ Extra-task-1/Extra-task-1/x64/Debug/Extra-task-1.log
@@ -1,0 +1,2 @@
+﻿  main.cpp
+  Extra-task-1.vcxproj -> C:\Projects\C++\Bonus1\cs211-extra-task-1\Extra-task-1\x64\Debug\Extra-task-1.exe
