+++ Extra-task-1/Extra-task-1/x64/Debug/Extra-task-1.tlog/Cl.items.tlog
@@ -1,0 +1,2 @@
+C:\Projects\C++\Bonus1\cs211-extra-task-1\Extra-task-1\Extra-task-1\extra-task-1.cpp;C:\Projects\C++\Bonus1\cs211-extra-task-1\Extra-task-1\Extra-task-1\x64\Debug\extra-task-1.obj
+C:\Projects\C++\Bonus1\cs211-extra-task-1\Extra-task-1\Extra-task-1\main.cpp;C:\Projects\C++\Bonus1\cs211-extra-task-1\Extra-task-1\Extra-task-1\x64\Debug\main.obj
