+++ Extra-task-1/Extra-task-1/x64/Debug/Extra-task-1.tlog/Extra-task-1.lastbuildstate
@@ -1,0 +1,2 @@
+PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.39.33519:TargetPlatformVersion=10.0.22621.0:
+Debug|x64|C:\Projects\C++\Bonus1\cs211-extra-task-1\Extra-task-1\|
