+++ Extra-task-1/Extra-task-1/x64/Debug/Extra-task-1.tlog/link.secondary.1.tlog
@@ -1,0 +1,2 @@
+^C:\PROJECTS\C++\BONUS1\CS211-EXTRA-TASK-1\EXTRA-TASK-1\EXTRA-TASK-1\X64\DEBUG\EXTRA-TASK-1.OBJ|C:\PROJECTS\C++\BONUS1\CS211-EXTRA-TASK-1\EXTRA-TASK-1\EXTRA-TASK-1\X64\DEBUG\MAIN.OBJ
+C:\Projects\C++\Bonus1\cs211-extra-task-1\Extra-task-1\Extra-task-1\x64\Debug\Extra-task-1.ilk
